// src/app/models/User.ts
export interface User {
    id?: number; // Optional for cases where the ID is generated by the backend
    firstName: string;
    lastName: string;
    email: string;
    password: string; 
    otpCode?: number; // Optional for cases where the OTP code is generated and managed by the backend
    verified?: boolean;
    role?: string;
   
    
}